Implemented @attr and List.of(Map.of()) xml patch structure
The code feels cumbersome and way to convoluted for what it does,
will try to see what Michal did in the Python version and improve on it.

Walking the trees should be easier than what I am doing now. Maybe
some preprocessing would help.

Kotlin may be beneficial here, may port the XML thing over
(data classes that generate .equals() automatically for values and so on)